// database/prisma-thinkthank/generated/zod/inputTypeSchemas/ReportScheduleUncheckedCreateInputSchema.ts
import type { Prisma } from '@prisma/client'

import { z } from 'zod'

export const ReportScheduleUncheckedCreateInputSchema: z.ZodType<Prisma.ReportScheduleUncheckedCreateInput> = z
    .object({
        id: z.string().cuid().optional(),
        reportId: z.string(),
        frequency: z.string(),
        dayOfWeek: z.number().int().optional().nullable(),
        dayOfMonth: z.number().int().optional().nullable(),
        hour: z.number().int(),
        minute: z.number().int(),
        isActive: z.boolean().optional(),
        lastRun: z.coerce.date().optional().nullable(),
        nextRun: z.coerce.date().optional().nullable()
    })
    .strict()

export default ReportScheduleUncheckedCreateInputSchema